add test for AE
