--- test/run_iae.py
@@ -1,79 +1,0 @@
-import numpy as np
-
-from qiskit import BasicAer
-from qiskit.aqua.algorithms import AmplitudeEstimationWithoutQPE
-from qiskit.aqua.algorithms import AmplitudeEstimation
-from qiskit.aqua.algorithms.single_sample.amplitude_estimation.q_factory import QFactory
-from qiskit.aqua.components.uncertainty_problems import UncertaintyProblem
-
-# the probability to be recovered
-probability = 0.3
-theta_p = 2 * np.arcsin(np.sqrt(probability))
-
-
-class BernoulliAFactory(UncertaintyProblem):
-    """
-    Circuit Factory representing the operator A.
-    A is used to initialize the state as well as to construct Q.
-    """
-
-    def __init__(self, probability=0.5):
-        #
-        super().__init__(1)
-        self._probability = probability
-        self.i_state = 0
-        self._theta_p = 2 * np.arcsin(np.sqrt(probability))
-
-    def build(self, qc, q, q_ancillas=None):
-        # A is a rotation of angle theta_p around the Y-axis
-        qc.ry(self._theta_p, q[self.i_state])
-
-
-class BernoulliQFactory(QFactory):
-    """
-    Circuit Factory representing the operator Q.
-    This implementation exploits the fact that powers of Q can be implemented efficiently by just multiplying the angle.
-    (amplitude estimation only requires controlled powers of Q, thus, only this method is overridden.)
-    """
-
-    def __init__(self, bernoulli_expected_value):
-        super().__init__(bernoulli_expected_value, i_objective=0)
-
-    def build(self, qc, q, q_ancillas=None):
-        i_state = self.a_factory.i_state
-        theta_p = self.a_factory._theta_p
-        # Q is a rotation of angle 2*theta_p around the Y-axis
-        qc.ry(2 * theta_p, q[i_state])
-
-    def build_power(self, qc, q, power, q_ancillas=None, use_basis_gates=True):
-        i_state = self.a_factory.i_state
-        theta_p = self.a_factory._theta_p
-        qc.ry(2 * power * theta_p, q[i_state])
-
-    def build_controlled_power(self, qc, q, q_control, power, q_ancillas=None, use_basis_gates=True):
-        i_state = self.a_factory.i_state
-        theta_p = self.a_factory._theta_p
-        qc.cry(2 * power * theta_p, q_control, q[i_state])
-
-
-# construct factories for A and Q
-bernoulli_a_factory = BernoulliAFactory(probability)
-bernoulli_q_factory = BernoulliQFactory(bernoulli_a_factory)
-
-
-# set number of evaluation qubits
-m = 3
-np.random.seed(11723)
-# construct amplitude estimation
-# here, we override the standard construction of Q since we know a more efficient way
-# (exploiting the fact that A and Q are just Y-rotations)
-ae = AmplitudeEstimationWithoutQPE(m, bernoulli_a_factory, i_objective=0, q_factory=bernoulli_q_factory)
-#ae = AmplitudeEstimation(m, bernoulli_a_factory, i_objective=0, q_factory=bernoulli_q_factory)
-
-
-shots = 10
-result = ae.run(quantum_instance=BasicAer.get_backend('qasm_simulator'), shots=shots)
-#result = ae.run(quantum_instance=BasicAer.get_backend('statevector_simulator'))
-
-for key, value in result.items():
-    print(key, value)
